sistema eventi amplificato

+++ src/models/carrier_assignment.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from src.database import Base
 
@@ -11,8 +11,8 @@ class CarrierAssignment(Base):
     postal_codes = Column(String(1000), nullable=True)  # JSON string with postal codes list
     countries = Column(String(1000), nullable=True)     # JSON string with country IDs list
     origin_carriers = Column(String(1000), nullable=True)  # JSON string with origin carrier IDs list
-    min_weight = Column(Float, nullable=True)
-    max_weight = Column(Float, nullable=True)
+    min_weight = Column(Numeric(10, 5), nullable=True)
+    max_weight = Column(Numeric(10, 5), nullable=True)
 
     # Relationship with CarrierApi
     carrier_api = relationship("CarrierApi", back_populates="carrier_assignments")
