+++ src/models/order_package.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Column, Float, ForeignKey
+from sqlalchemy import Integer, Column, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from src import Base
 
@@ -9,12 +9,12 @@ class OrderPackage(Base):
     id_order_package = Column(Integer, primary_key=True, index=True)
     id_order = Column(Integer, ForeignKey('orders.id_order'), index=True, nullable=True, default=None)
     id_order_document = Column(Integer, ForeignKey('orders_document.id_order_document'), index=True, nullable=True, default=None)
-    height = Column(Float)
-    width = Column(Float)
-    depth = Column(Float)
-    length = Column(Float)
-    weight = Column(Float)
-    value = Column(Float, default=0.0)
+    height = Column(Numeric(10, 5))
+    width = Column(Numeric(10, 5))
+    depth = Column(Numeric(10, 5))
+    length = Column(Numeric(10, 5))
+    weight = Column(Numeric(10, 5))
+    value = Column(Numeric(10, 5), default=0.0)
     
     # Relazioni
     order = relationship("Order", back_populates="order_packages")
