+++ src/models/shipping.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, String, Date, Float, Text
+from sqlalchemy import Column, Integer, String, Date, Numeric, Text
 
 from ..database import Base
 
@@ -14,9 +14,9 @@ class Shipping(Base):
     id_shipping_state = Column(Integer, index=True, default=None)
     id_tax = Column(Integer)
     tracking = Column(String(100), default=None, index=True)
-    weight = Column(Float, default=0)
-    price_tax_incl = Column(Float, default=0)
-    price_tax_excl = Column(Float, default=0)
+    weight = Column(Numeric(10, 5), default=0)
+    price_tax_incl = Column(Numeric(10, 5), default=0)
+    price_tax_excl = Column(Numeric(10, 5), default=0)
     shipping_message = Column(Text)
     date_add = Column(Date, default=datetime.today)
